add custom validation reg exp for phone number

+++ pages/bucket.vue
@@ -15,13 +15,16 @@
     </div>
     <form class="row" novalidate @submit.prevent="makeOrder">
       <div class="col-12 form-input-wrapper">
-        <input class="form-control" type="text" name="name" placeholder="Введите имя и фамилию" v-model="name">
+        <input class="form-control" type="text" v-validate:name="'required'" name="name" placeholder="Введите имя и фамилию" v-model="name">
+        <span class="error-default" v-show="errors.has('name')"> {{errors.first('name')}} </span>
       </div>
       <div class="col-12 form-input-wrapper">
-        <input class="form-control" type="text" name="email" placeholder="Введите эл. почту" v-model="email">
+        <input class="form-control" type="text" name="email" v-validate:email="'required|email'" placeholder="Введите эл. почту" v-model="email">
+        <span class="error-default" v-show="errors.has('email')"> {{errors.first('email')}} </span>
       </div>
       <div class="col-12 form-input-wrapper">
-        <input class="form-control" type="number" name="tel" placeholder="Введите номер телефона" v-model="tel">
+        <input class="form-control" type="number" v-validate:tel="'required|phoneNumber'" name="tel" placeholder="Введите номер телефона" v-model="tel">
+        <span class="error-default" v-show="errors.has('tel')"> {{errors.first('tel')}} </span>
       </div>
       <div class="col-12 form-input-wrapper">
         <textarea class="form-control" type="text" name="comment" placeholder="Добавьте пожелания к заказу, если таковые имеются" v-model="comment"></textarea>
@@ -59,17 +62,20 @@ export default {
   },
   methods: {
     async makeOrder() {
-      await axios.post("/api/orders", {
-        name: this.name,
-        email: this.email,
-        products: this.products,
-        tel: this.tel,
-        comment: this.comment,
-        totalPrice: this.totalBucketPrice
-      });
-      storeService.removeCookie('mrbucket');
-      storeService.cleanBucket(this.$store);
-      alert("Заказ оформлен");
+      let valid = await this.$validator.validateAll();
+      if (valid) {
+        await axios.post("/api/orders", {
+          name: this.name,
+          email: this.email,
+          products: this.products,
+          tel: this.tel,
+          comment: this.comment,
+          totalPrice: this.totalBucketPrice
+        });
+        storeService.removeCookie("mrbucket");
+        storeService.cleanBucket(this.$store);
+        alert("Заказ оформлен");
+      }
     }
   }
 };
